feat: move account type & constant defs to shared location
