--- apps/account-api/src/controllers/index.ts
@@ -1,2 +1,0 @@
-export * from './health.controller';
-export * from './v1';
