--- apps/account-api/src/controllers/v1/index.ts
@@ -1,4 +1,0 @@
-export * from './accounts-v1.controller';
-export * from './delegation-v1.controller';
-export * from './handles-v1.controller';
-export * from './keys-v1.controller';
