--- apps/account-api/src/controllers/v1/keys-v1.controller.ts
@@ -1,79 +1,0 @@
-import { KeysService } from '#account-api/services/keys.service';
-import { EnqueueService } from '#account-lib/services/enqueue-request.service';
-import { TransactionType } from '#account-lib/types/enums';
-import {
-  Controller,
-  Get,
-  HttpCode,
-  HttpStatus,
-  Logger,
-  Param,
-  HttpException,
-  Body,
-  Post,
-  UseGuards,
-} from '@nestjs/common';
-import { ApiBody, ApiOkResponse, ApiOperation, ApiTags } from '@nestjs/swagger';
-import { KeysRequestDto, AddKeyRequestDto } from '#account-lib/types/dtos/keys.request.dto';
-import { TransactionResponse } from '#account-lib/types/dtos/transaction.response.dto';
-import { KeysResponse } from '#account-lib/types/dtos/keys.response.dto';
-import { ReadOnlyGuard } from '#account-api/guards/read-only.guard';
-
-@Controller('v1/keys')
-@ApiTags('v1/keys')
-@UseGuards(ReadOnlyGuard) // Apply guard at the controller level
-export class KeysControllerV1 {
-  private readonly logger: Logger;
-
-  constructor(
-    private keysService: KeysService,
-    private enqueueService: EnqueueService,
-  ) {
-    this.logger = new Logger(this.constructor.name);
-  }
-
-  @Post('add')
-  @HttpCode(HttpStatus.OK)
-  @ApiOperation({ summary: 'Add new control keys for an MSA Id' })
-  @ApiOkResponse({ description: 'Found public keys' })
-  @ApiBody({ type: KeysRequestDto })
-  /**
-   * Add new control keys for an MSA Id.
-   * @param queryParams - The query parameters for adding the public keys.
-   * @returns A promise that resolves to an array of public keys associated with the given msaId.
-   * @throws An error if no public keys can be found.
-   */
-  async addKey(@Body() addKeysRequest: KeysRequestDto): Promise<TransactionResponse> {
-    try {
-      const response = await this.enqueueService.enqueueRequest<AddKeyRequestDto>({
-        ...addKeysRequest,
-        type: TransactionType.ADD_KEY,
-      });
-      this.logger.log(`AddKey in progress. referenceId: ${response.referenceId}`);
-      return response;
-    } catch (error) {
-      this.logger.error(error);
-      throw new HttpException('Failed to find public keys for the given MSA Id', HttpStatus.BAD_REQUEST);
-    }
-  }
-
-  @Get(':msaId')
-  @HttpCode(HttpStatus.OK)
-  @ApiOperation({ summary: 'Fetch public keys given an MSA Id' })
-  @ApiOkResponse({ description: 'Found public keys' })
-  /**
-   * Gets public keys.
-   * @param queryParams - The query parameters for getting the public keys.
-   * @returns A promise that resolves to an array of public keys associated with the given msaId.
-   * @throws An error if no public keys can be found.
-   */
-  async getKeys(@Param('msaId') msaId: string): Promise<KeysResponse> {
-    try {
-      const keys = await this.keysService.getKeysByMsa(msaId);
-      return keys;
-    } catch (error) {
-      this.logger.error(error);
-      throw new HttpException('Failed to find public keys for the given msaId', HttpStatus.BAD_REQUEST);
-    }
-  }
-}
