--- apps/account-api/src/generate-metadata.ts
@@ -1,11 +1,0 @@
-// eslint-disable-next-line import/no-extraneous-dependencies
-import { PluginMetadataGenerator } from '@nestjs/cli/lib/compiler/plugins';
-import { ReadonlyVisitor } from '@nestjs/swagger/dist/plugin';
-
-const generator = new PluginMetadataGenerator();
-generator.generate({
-  visitors: [new ReadonlyVisitor({ introspectComments: true, pathToSource: __dirname })],
-  outputDir: __dirname,
-  watch: false,
-  tsconfigPath: 'apps/account-api/tsconfig.app.json',
-});
