--- apps/account-api/src/services/delegation.service.ts
@@ -1,39 +1,0 @@
-import { BlockchainService } from '#account-lib/blockchain/blockchain.service';
-import { ConfigService } from '#account-lib/config/config.service';
-import { DelegationResponse } from '#account-lib/types/dtos/delegation.response.dto';
-import { Injectable, Logger } from '@nestjs/common';
-
-@Injectable()
-export class DelegationService {
-  private readonly logger: Logger;
-
-  constructor(
-    private configService: ConfigService,
-    private blockchainService: BlockchainService,
-  ) {
-    this.logger = new Logger(this.constructor.name);
-  }
-
-  async getDelegation(msaId: string): Promise<DelegationResponse> {
-    const isValidMsaId = await this.blockchainService.isValidMsaId(msaId);
-    if (isValidMsaId) {
-      const { providerId } = this.configService;
-
-      const commonPrimitivesMsaDelegation = await this.blockchainService.getCommonPrimitivesMsaDelegation(
-        msaId,
-        providerId,
-      );
-
-      if (commonPrimitivesMsaDelegation) {
-        const delegationResponse: DelegationResponse = {
-          providerId,
-          schemaPermissions: commonPrimitivesMsaDelegation.schemaPermissions,
-          revokedAt: commonPrimitivesMsaDelegation.revokedAt,
-        };
-        return delegationResponse;
-      }
-      throw new Error('Failed to find the delegation.');
-    }
-    throw new Error('Invalid msaId.');
-  }
-}
