--- apps/account-api/src/services/index.ts
@@ -1,4 +1,0 @@
-export * from './accounts.service';
-export * from './delegation.service';
-export * from './handles.service';
-export * from './keys.service';
