+++ apps/account-worker/src/transaction_publisher/publisher.service.ts
@@ -11,12 +11,11 @@ import { SchedulerRegistry } from '@nestjs/schedule';
 import { BlockchainService, ICapacityInfo } from '#account-lib/blockchain/blockchain.service';
 import { createKeys } from '#account-lib/blockchain/create-keys';
 import { NonceService } from '#account-lib/services/nonce.service';
-import { TransactionType } from '#account-lib/types/enums';
-import { QueueConstants } from '#account-lib/queues';
+import { TransactionType } from '#types/enums/account-enums';
+import * as QueueConstants from '#types/constants/queue.constants';
 import { BaseConsumer } from '#account-worker/BaseConsumer';
-import { RedisUtils, TransactionData } from '#account-lib';
 import { ConfigService } from '#account-lib/config/config.service';
-import { ITxStatus } from '#account-lib/interfaces/tx-status.interface';
+import { ITxStatus } from '#types/interfaces/tx-status.interface';
 import { HexString } from '@polkadot/util/types';
 import {
   CAPACITY_AVAILABLE_EVENT,
@@ -24,6 +23,8 @@ import {
   CapacityCheckerService,
 } from '#account-lib/blockchain/capacity-checker.service';
 import { OnEvent } from '@nestjs/event-emitter';
+import { TransactionData } from '#types/dtos/account';
+import { RedisUtils } from '#account-lib/utils/redis';
 
 export const SECONDS_PER_BLOCK = 12;
 const CAPACITY_EPOCH_TIMEOUT_NAME = 'capacity_check';
