+++ jest.config.json
@@ -34,6 +34,7 @@
     "^#account-lib(|/.*)$": "<rootDir>/libs/account-lib/src/$1",
     "^#content-publishing-lib(|/.*)$": "<rootDir>/libs/content-publishing-lib/src/$1",
     "^#content-watcher-lib(|/.*)$": "<rootDir>/libs/content-watcher-lib/src/$1",
-    "^#graph-lib(|/.*)$": "<rootDir>/libs/graph-lib/src/$1"
+    "^#graph-lib(|/.*)$": "<rootDir>/libs/graph-lib/src/$1",
+    "^#types(|/.*)$": "<rootDir>/libs/types/src/$1"
   }
 }
