+++ libs/account-lib/src/blockchain/blockchain.service.ts
@@ -8,7 +8,6 @@ import { SubmittableExtrinsic } from '@polkadot/api/types';
 import { AnyNumber, ISubmittableResult } from '@polkadot/types/types';
 import { Bytes, Option, u32 } from '@polkadot/types';
 import {
-  CommonPrimitivesHandlesClaimHandlePayload,
   CommonPrimitivesMsaDelegation,
   FrameSystemEventRecord,
   PalletCapacityCapacityDetails,
@@ -17,13 +16,13 @@ import {
 } from '@polkadot/types/lookup';
 import { HandleResponse, KeyInfoResponse } from '@frequency-chain/api-augment/interfaces';
 import { ConfigService } from '#account-lib/config/config.service';
-import { TransactionType } from '#account-lib/types/enums';
+import { TransactionType } from '#types/enums/account-enums';
 import { HexString } from '@polkadot/util/types';
 import { decodeAddress } from '@polkadot/util-crypto';
-import { KeysRequestDto } from '#account-lib/types/dtos/keys.request.dto';
-import { PublishHandleRequestDto } from '#account-lib/types/dtos/handles.request.dto';
-import { TransactionData } from '#account-lib/types/dtos/transaction.request.dto';
-import { HandleResponseDto } from '#account-lib/types/dtos/accounts.response.dto';
+import { KeysRequestDto } from '#types/dtos/account/keys.request.dto';
+import { PublishHandleRequestDto } from '#types/dtos/account/handles.request.dto';
+import { TransactionData } from '#types/dtos/account/transaction.request.dto';
+import { HandleResponseDto } from '#types/dtos/account/accounts.response.dto';
 import { Extrinsic } from './extrinsic';
 
 export type Sr25519Signature = { Sr25519: HexString };
