+++ libs/account-lib/src/queues/index.ts
@@ -1,2 +1,1 @@
-export * as QueueConstants from './queue.constants';
 export * from './queues.module';
