+++ libs/account-lib/src/services/enqueue-request.service.ts
@@ -5,7 +5,7 @@ import { Queue } from 'bullmq';
 import { createHash } from 'crypto';
 import { QueueConstants } from '#account-lib/queues';
 import { ConfigService } from '#account-lib/config/config.service';
-import { TransactionResponse, TransactionData } from '#account-lib/types/dtos';
+import { TransactionResponse, TransactionData } from '#types/dtos/account';
 
 @Injectable()
 export class EnqueueService {
