+++ libs/types/src/constants/blockchain-constants.ts
@@ -1,0 +1,3 @@
+export namespace BlockchainConstants {
+  export const SECONDS_PER_BLOCK = 12;
+}
