+++ libs/types/src/dtos/account/handles.request.dto.ts
@@ -2,7 +2,7 @@
 import { IsNotEmpty } from 'class-validator';
 import { ApiProperty } from '@nestjs/swagger';
 import { HexString } from '@polkadot/util/types';
-import { TransactionType } from '../enums';
+import { TransactionType } from '../../enums/account-enums';
 
 class HandlePayloadDto {
   @ApiProperty()
