+++ libs/types/src/dtos/account/transaction.request.dto.ts
@@ -1,6 +1,6 @@
 import { BlockHash } from '@polkadot/types/interfaces';
 import { HexString } from '@polkadot/util/types';
-import { PublishHandleRequestDto } from './handles.request.dto';
+import { PublishHandleRequestDto } from '#types/dtos/account/handles.request.dto';
 import { PublishSIWFSignupRequestDto } from './wallet.login.request.dto';
 import { PublishKeysRequestDto } from './keys.request.dto';
