+++ libs/types/src/interfaces/capacity-limit.interface.ts
