+++ libs/types/src/interfaces/index.ts
@@ -1,0 +1,2 @@
+export * from './capacity-limit.interface';
+export * from './tx-status.interface';
