+++ libs/types/src/interfaces/tx-status.interface.ts
@@ -1,4 +1,4 @@
-import { TxMonitorJob } from '#account-lib/types/dtos/transaction.request.dto';
+import { TxMonitorJob } from '#types/dtos/account/transaction.request.dto';
 
 export interface ITxStatus extends Pick<TxMonitorJob, 'providerId' | 'referenceId' | 'txHash' | 'type'> {
   successEvent: {
