+++ nest-cli.json
@@ -108,6 +108,15 @@
       "compilerOptions": {
         "tsConfigPath": "apps/graph-worker/tsconfig.app.json"
       }
+    },
+    "types": {
+      "type": "library",
+      "root": "libs/types",
+      "entryFile": "index",
+      "sourceRoot": "libs/types/src",
+      "compilerOptions": {
+        "tsConfigPath": "libs/types/tsconfig.lib.json"
+      }
     }
   }
 }
